scara/requirements.txt
web3==6.11.1
python-dotenv==1.0.0
networkx==3.2.1
aiohttp==3.9.1
asyncio==3.4.3
eth-account==0.10.0
eth-typing==4.0.0
eth-utils==2.3.1
openai==1.3.7
plotly==5.18.0
pandas==2.1.4
numpy==1.26.2
pytest==7.4.3
pytest-asyncio==0.23.2 